Refactor new ops after rebase
