+++ lighthouse/dialects/transform/transform_ext/__init__.py
@@ -5,10 +5,19 @@ from .ops.wrap_in_benching_func import wrap_in_benching_func
 from .ops.get_named_attribute import get_named_attribute
 from .ops.param_cmp_eq import param_cmp_eq
 from .ops.replace import replace
+from .ops.convert_func_results_to_args import convert_func_results_to_args
+from .ops.extract_handle import extract_handle
+from .ops.get_tileable_consumers import get_tileable_consumers
+from .ops.get_tiling_sizes import get_tiling_sizes
 
 __all__ = [
     "TransformExtensionDialect",
+    "convert_func_results_to_args",
+    "extract_handle",
     "get_named_attribute",
+    "get_named_attribute",
+    "get_tileable_consumers",
+    "get_tiling_sizes",
     "param_cmp_eq",
     "register_and_load",
     "replace",
