+++ lighthouse/transform/x86/cache_tiling.py
@@ -2,7 +2,7 @@ from mlir import ir
 from mlir.dialects import transform
 from mlir.dialects.transform import structured
 
-from lighthouse.dialects import transform_ext
+from lighthouse.dialects.transform import transform_ext
 
 
 def matmul_cache_tiling(
